requirements.txt
flask
flask-restx
gunicorn
chromadb
numpy
pandas
python-dotenv
orjson
rapidfuzz
twilio
openai
pydantic
Flask-Session
undetected-chromedriver 
selenium 
webdriver-manager 
beautifulsoup4 
lxml
pinecone
psycopg2-binary
pillow
redis